refactor: 객체 이름 바꾸기
