+++ src/main/java/janggi/unit/SoldierUnitRule.java
@@ -5,7 +5,7 @@ import janggi.position.Route;
 import java.util.ArrayList;
 import java.util.List;
 
-public class JolUnitRule implements UnitRule {
+public class SoldierUnitRule implements UnitRule {
     @Override
     public List<Route> calculateAllRoute(Position start) {
         List<Route> routes = new ArrayList<>();
